Fonction loadSave terminée + carnet de bord
loadSave charge tous les paramètres (normalement).
J'ai changé quelques trucs sur les classes (surtout au niveau des types
de return ou de paramètres) faudra penser à le répercuter sur le
diagramme.
-> tous les changements importants par rapport au diagramme on le met
dans le carnet de bord. On peut aussi noter ce qu'il faudra penser à
faire plus tard.

+++ headers/System.h
@@ -24,7 +24,7 @@ class System
 	GUI *interface;
 	string location;
 	string saveName;
-	int timeSpeed;
+	float timeSpeed;
 	string weather;
 	bool endGame;
 	bool endProgram;
